refactor(flui-layer): Mythos Step 7 -- extract LayerTree tests to tests/
Move the 20-test inline `#[cfg(test)] mod tests { ... }` block from
`crates/flui-layer/src/tree/layer_tree.rs` to a new
`crates/flui-layer/tests/layer_tree.rs` integration test file. The
production module shrinks to ~315 LOC focused on storage logic
(LayerNode + LayerTree).

LOC delta:

- `src/tree/layer_tree.rs`: 876 -> 543 LOC (-333 LOC of inline tests +
  unused imports cleanup).
- `tests/layer_tree.rs`: new, 330 LOC (matches the extracted block plus
  the integration-test preamble: `use flui_layer::*; use flui_foundation;
  use flui_types::*`).
- Net: -3 LOC (the integration-test file is slightly larger than the
  extracted block due to the module docstring and the integration-style
  imports replacing `super::*` + `crate::layer::CanvasLayer`).

Test scenarios preserved verbatim. All 20 tests pass after the move:

- test_layer_tree_new / with_capacity / insert / get / get_layer / remove
- test_layer_tree_parent_child / remove_child / set_root / clear / iter
- test_layer_node_with_element_id / with_offset / needs_compositing
- test_clear_children
- test_append_layer / multiple_times / bulk / empty
- test_layer_composition_integration

Also trimmed the production module's `use` statements that the deleted
tests previously needed (`Matrix4`, `RRect`, `Rect`, `Clip`, `Path` --
all consumed only by `LayerTree::push_*` helpers deleted in Step 5 and
their tests deleted here).

Production module structure after this step:

- Module docstring (~5 LOC)
- imports (~5 LOC)
- LayerNode struct + impl (~170 LOC)
- LayerTree struct + impl (Slab arena, root/insert/get/remove/parent/
  children/append_layer/append_layers/clear_children + iteration) (~360 LOC)
- impl Default for LayerTree (~5 LOC)

Files touched:

- Create: `crates/flui-layer/tests/layer_tree.rs` -- the integration test
  file with module-level documentation explaining the extraction.
- Modify: `crates/flui-layer/src/tree/layer_tree.rs`:
  * Delete the `#[cfg(test)] mod tests { ... }` block (lines 549-876).
  * Trim `use` block to drop now-unused imports
    (`Matrix4`, `RRect`, `Rect`, `Clip`, `Path`).
  * Section header `// ========== TESTS ==========` deleted.

Verification:

- `cargo check --workspace`: clean.
- `cargo test -p flui-layer --lib`: 243 passed, 0 failed (down from 263;
  the 20 moved tests are now integration tests, not lib tests).
- `cargo test -p flui-layer --tests`: 20 passed, 0 failed
  (the `--test layer_tree` binary).
- `cargo test -p flui-layer --tests damage_tracking`: 8 passed
  (untouched).

Pre-existing doctest failures noted: `cargo test -p flui-layer --doc`
reports 21 failures across `compositor.rs`, `scene.rs`, and various
`layer/*.rs` doc-examples that use `Offset::new(100.0, 50.0)` where
`Pixels` wrapping is now required. These failures pre-date the Mythos
chain (commits da54fe80 and earlier touched the same doc-comments
verbatim; no chain step modified the doctest content semantically) and
are tracked as a separate concern in the per-crate `ARCHITECTURE.md`
Friction log (Mythos Step 12).

Plan: docs/plans/2026-05-20-002-feat-flui-layer-mythos-redesign-plan.md U7.
